common LangChainClient class for BAM and Ollama
Both BAMClient and OllamaClient now inherite from the new LangChainClient class.
This to avoid the confusing situation where OllamaClient was depending on bam.py.
