+++ pyproject.toml
@@ -82,14 +82,15 @@ profile = "black"
 [tool.pyright]
 include = [
   "ansible_wisdom/ai/api/aws/wca_secret_manager.py",
-  "ansible_wisdom/users/authz_checker.py",
   "ansible_wisdom/ai/api/model_client/bam_client.py",
   "ansible_wisdom/ai/api/model_client/dummy_client.py",
   "ansible_wisdom/ai/api/model_client/grpc_client.py",
   "ansible_wisdom/ai/api/model_client/http_client.py",
+  "ansible_wisdom/ai/api/model_client/langchain.py",
   "ansible_wisdom/ai/api/model_client/llamacpp_client.py",
   "ansible_wisdom/ai/api/model_client/ollama_client.py",
-  "ansible_wisdom/ai/api/model_client/wca_client.py"
+  "ansible_wisdom/ai/api/model_client/wca_client.py",
+  "ansible_wisdom/users/authz_checker.py"
 ]
 exclude = ["**/test_*.py", "ansible_wisdom/*/migrations/*.py"]
